pyproject.toml
[project]
name = "base-repo-structure"
version = "0.1.0"
description = "Production-grade FastAPI boilerplate with Google ADK multi-agent architecture"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.38.0",
    "google-adk>=1.19.0",
    "mcp>=1.23.1",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "loguru>=0.7.3",
    "modern-di>=1.1.0",
    "modern-di-fastapi>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
